Improved reporting a little
We see a number of valid responses which are instantly returned
but I don't at this stage know why. It seems like any delay that's occurring
might be to do with the web server rather than our explicit delay because
we see more of them if randomise the request starts a little.